backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Clients
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# AI and ML
langchain==0.1.17
langchain-core==0.1.52
langchain-google-genai==1.0.6
langchain-community==0.0.38
langchain-ollama==0.1.0
google-generativeai==0.5.4
ollama==0.2.1
scikit-learn==1.4.2
transformers==4.40.0
torch==2.3.0
sentence-transformers==2.7.0
# Additional AI packages
joblib==1.3.2
psutil==5.9.6
GPUtil==1.4.0
scipy==1.11.4

# Data Processing
pandas==2.1.4
numpy==1.25.2
python-dateutil==2.8.2

# Task Queue (for background jobs)
celery==5.3.4
redis==5.0.1

# Scheduling
APScheduler==3.10.4

# Logging and Monitoring
loguru==0.7.2
sentry-sdk[fastapi]==1.38.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Marketplace APIs (will need specific SDKs)
selenium==4.15.2
beautifulsoup4==4.12.2

# File handling
openpyxl==3.1.2
Pillow==10.1.0

# Caching
aiocache==0.12.2

# Validation and Serialization
marshmallow==3.20.1
orjson==3.9.10

# API Integration and Retry Logic
tenacity==8.2.3

# XML Processing (for 11st API)
lxml==4.9.3

# Trend Analysis
pytrends==4.9.2